+++ cloudscale/config.go
@@ -2,6 +2,7 @@ package cloudscale
 
 import (
 	"github.com/cloudscale-ch/cloudscale-go-sdk"
+	"github.com/hashicorp/terraform/helper/logging"
 	"golang.org/x/oauth2"
 )
 
@@ -10,10 +11,12 @@ type Config struct {
 }
 
 func (c *Config) Client() (*cloudscale.Client, error) {
-
 	tc := oauth2.NewClient(oauth2.NoContext, oauth2.StaticTokenSource(
 		&oauth2.Token{AccessToken: c.Token},
 	))
+
+	tc.Transport = logging.NewTransport("Cloudscale", tc.Transport)
+
 	client := cloudscale.NewClient(tc)
 
 	return client, nil