+++ cloudscale/resource_cloudscale_server_test.go
@@ -14,6 +14,8 @@ import (
 	"github.com/hashicorp/terraform/terraform"
 )
 
+const DefaultImageSlug = "debian-9"
+
 func init() {
 	resource.AddTestSweepers("cloudscale_server", &resource.Sweeper{
 		Name: "cloudscale_server",
@@ -67,9 +69,10 @@ func TestAccCloudscaleServer_Basic(t *testing.T) {
 					resource.TestCheckResourceAttr(
 						"cloudscale_server.basic", "flavor_slug", "flex-2"),
 					resource.TestCheckResourceAttr(
-						"cloudscale_server.basic", "image_slug", "debian-8"),
+						"cloudscale_server.basic", "image_slug", DefaultImageSlug),
 					resource.TestCheckResourceAttr(
 						"cloudscale_server.basic", "interfaces.0.type", "public"),
+					testAccCheckServerIp("cloudscale_server.basic"),
 				),
 			},
 		},
@@ -96,7 +99,7 @@ func TestAccCloudscaleServer_Basic_stopped(t *testing.T) {
 					resource.TestCheckResourceAttr(
 						"cloudscale_server.basic", "flavor_slug", "flex-2"),
 					resource.TestCheckResourceAttr(
-						"cloudscale_server.basic", "image_slug", "debian-8"),
+						"cloudscale_server.basic", "image_slug", DefaultImageSlug),
 					resource.TestCheckResourceAttr(
 						"cloudscale_server.basic", "status", "stopped"),
 				),
@@ -148,7 +151,8 @@ func TestAccCloudscaleServer_Update(t *testing.T) {
 					resource.TestCheckResourceAttr(
 						"cloudscale_server.basic", "flavor_slug", "flex-2"),
 					resource.TestCheckResourceAttr(
-						"cloudscale_server.basic", "image_slug", "debian-8"),
+						"cloudscale_server.basic", "image_slug", DefaultImageSlug),
+					testAccCheckServerIp("cloudscale_server.basic"),
 				),
 			},
 			{
@@ -170,6 +174,7 @@ func TestAccCloudscaleServer_Update(t *testing.T) {
 						"cloudscale_server.basic", "name", fmt.Sprintf("terraform-%d", rInt)),
 					resource.TestCheckResourceAttr(
 						"cloudscale_server.basic", "status", "running"),
+					testAccCheckServerIp("cloudscale_server.basic"),
 					testAccCheckServerChanged(t, &afterCreate, &afterUpdate),
 				),
 			},
@@ -197,7 +202,8 @@ func TestAccCloudscaleServer_Recreated(t *testing.T) {
 					resource.TestCheckResourceAttr(
 						"cloudscale_server.basic", "flavor_slug", "flex-2"),
 					resource.TestCheckResourceAttr(
-						"cloudscale_server.basic", "image_slug", "debian-8"),
+						"cloudscale_server.basic", "image_slug", DefaultImageSlug),
+					testAccCheckServerIp("cloudscale_server.basic"),
 				),
 			},
 			{
@@ -210,6 +216,7 @@ func TestAccCloudscaleServer_Recreated(t *testing.T) {
 						"cloudscale_server.basic", "flavor_slug", "flex-4"),
 					resource.TestCheckResourceAttr(
 						"cloudscale_server.basic", "interfaces.#", "2"),
+					testAccCheckServerIp("cloudscale_server.basic"),
 					testAccCheckServerRecreated(t, &afterCreate, &afterUpdate),
 				),
 			},
@@ -234,6 +241,7 @@ func TestAccCloudscaleServer_PrivateNetwork(t *testing.T) {
 						"cloudscale_server.private", "interfaces.#", "1"),
 					resource.TestCheckResourceAttr(
 						"cloudscale_server.private", "interfaces.0.type", "private"),
+					testAccCheckServerIp("cloudscale_server.private"),
 				),
 			},
 		},
@@ -303,7 +311,7 @@ func testAccCheckCloudscaleServerExists(n string, server *cloudscale.Server) res
 func testAccCheckCloudscaleServerAttributes(server *cloudscale.Server) resource.TestCheckFunc {
 	return func(s *terraform.State) error {
 
-		if server.Image.Slug != "debian-8" {
+		if server.Image.Slug != DefaultImageSlug {
 			return fmt.Errorf("Bad image_slug_slug: %s", server.Image.Slug)
 		}
 
@@ -315,6 +323,56 @@ func testAccCheckCloudscaleServerAttributes(server *cloudscale.Server) resource.
 			return fmt.Errorf("Bad volumes_size_gb: %d", server.Volumes[0].SizeGB)
 		}
 
+		return nil
+	}
+}
+
+func testAccCheckServerIp(n string) resource.TestCheckFunc {
+	return func(s *terraform.State) error {
+		rs, ok := s.RootModule().Resources[n]
+		if !ok {
+			return fmt.Errorf("Not found: %s", n)
+		}
+
+		if rs.Primary.ID == "" {
+			return fmt.Errorf("No Server ID is set")
+		}
+
+		client := testAccProvider.Meta().(*cloudscale.Client)
+
+		id := rs.Primary.ID
+
+		// Try to find the server
+		retrieveServer, err := client.Servers.Get(context.Background(), id)
+
+		if err != nil {
+			return err
+		}
+
+		if retrieveServer.UUID != rs.Primary.ID {
+			return fmt.Errorf("Server not found")
+		}
+
+		for _, networkInterface := range retrieveServer.Interfaces {
+			for _, ipAddress := range networkInterface.Adresses {
+				if ipAddress.Version == 4 && networkInterface.Type == "public" {
+					err := resource.TestCheckResourceAttr(n, "public_ipv4_address", ipAddress.Address)(s)
+					if err != nil {
+						return err
+					}
+				} else if ipAddress.Version == 4 && networkInterface.Type == "private" {
+					err := resource.TestCheckResourceAttr(n, "private_ipv4_address", ipAddress.Address)(s)
+					if err != nil {
+						return err
+					}
+				} else if ipAddress.Version == 6 && networkInterface.Type == "public" {
+					err := resource.TestCheckResourceAttr(n, "public_ipv6_address", ipAddress.Address)(s)
+					if err != nil {
+						return err
+					}
+				}
+			}
+		}
 		return nil
 	}
 }
@@ -358,10 +416,10 @@ func testAccCheckCloudscaleServerConfig_basic(rInt int) string {
 resource "cloudscale_server" "basic" {
   name      					= "terraform-%d"
   flavor_slug    			= "flex-2"
-  image_slug     			= "debian-8"
+  image_slug     			= "%s"
   volume_size_gb			= 10
   ssh_keys 						= ["ecdsa-sha2-nistp256 AAAAE2VjZHNhLXNoYTItbmlzdHAyNTYAAAAIbmlzdHAyNTYAAABBBFEepRNW5hDct4AdJ8oYsb4lNP5E9XY5fnz3ZvgNCEv7m48+bhUjJXUPuamWix3zigp2lgJHC6SChI/okJ41GUY=", "ecdsa-sha2-nistp256 AAAAE2VjZHNhLXNoYTItbmlzdHAyNTYAAAAIbmlzdHAyNTYAAABBBFEepRNW5hDct4AdJ8oYsb4lNP5E9XY5fnz3ZvgNCEv7m48+bhUjJXUPuamWix3zigp2lgJHC6SChI/okJ41GUY="]
-}`, rInt)
+}`, rInt, DefaultImageSlug)
 }
 
 func testAccCheckCloudscaleServerConfig_basic_stopped(rInt int) string {
@@ -369,11 +427,11 @@ func testAccCheckCloudscaleServerConfig_basic_stopped(rInt int) string {
 resource "cloudscale_server" "basic" {
   name      					= "terraform-%d"
   flavor_slug    			= "flex-2"
-  image_slug     			= "debian-8"
+  image_slug     			= "%s"
   volume_size_gb			= 10
   status							= "stopped"
   ssh_keys 						= ["ecdsa-sha2-nistp256 AAAAE2VjZHNhLXNoYTItbmlzdHAyNTYAAAAIbmlzdHAyNTYAAABBBFEepRNW5hDct4AdJ8oYsb4lNP5E9XY5fnz3ZvgNCEv7m48+bhUjJXUPuamWix3zigp2lgJHC6SChI/okJ41GUY="]
-}`, rInt)
+}`, rInt, DefaultImageSlug)
 }
 
 func testAccCheckCloudscaleServerConfig_update_state_stopped(rInt int) string {
@@ -381,11 +439,11 @@ func testAccCheckCloudscaleServerConfig_update_state_stopped(rInt int) string {
 resource "cloudscale_server" "basic" {
   name      					= "terraform-%d"
   flavor_slug    			= "flex-2"
-  image_slug     			= "debian-8"
+  image_slug     			= "%s"
   volume_size_gb			= 10
 	status 							= "stopped"
   ssh_keys 						= ["ecdsa-sha2-nistp256 AAAAE2VjZHNhLXNoYTItbmlzdHAyNTYAAAAIbmlzdHAyNTYAAABBBFEepRNW5hDct4AdJ8oYsb4lNP5E9XY5fnz3ZvgNCEv7m48+bhUjJXUPuamWix3zigp2lgJHC6SChI/okJ41GUY="]
-}`, rInt)
+}`, rInt, DefaultImageSlug)
 }
 
 func testAccCheckCloudscaleServerConfig_anti_affinity_group(aInt, bInt int) string {
@@ -393,16 +451,16 @@ func testAccCheckCloudscaleServerConfig_anti_affinity_group(aInt, bInt int) stri
 resource "cloudscale_server" "dbmaster" {
   name      					= "terraform-%d"
   flavor_slug    			= "flex-2"
-  image_slug     			= "debian-8"
+  image_slug     			= "%s"
   ssh_keys 						= ["ecdsa-sha2-nistp256 AAAAE2VjZHNhLXNoYTItbmlzdHAyNTYAAAAIbmlzdHAyNTYAAABBBFEepRNW5hDct4AdJ8oYsb4lNP5E9XY5fnz3ZvgNCEv7m48+bhUjJXUPuamWix3zigp2lgJHC6SChI/okJ41GUY="]
 }
 resource "cloudscale_server" "web" {
   name      					= "terraform-%d"
   flavor_slug    			= "flex-2"
-  image_slug     			= "debian-8"
+  image_slug     			= "%s"
   ssh_keys 						= ["ecdsa-sha2-nistp256 AAAAE2VjZHNhLXNoYTItbmlzdHAyNTYAAAAIbmlzdHAyNTYAAABBBFEepRNW5hDct4AdJ8oYsb4lNP5E9XY5fnz3ZvgNCEv7m48+bhUjJXUPuamWix3zigp2lgJHC6SChI/okJ41GUY="]
 	anti_affinity_uuid 	= "${cloudscale_server.dbmaster.id}"
-}`, aInt, bInt)
+}`, aInt, DefaultImageSlug, bInt, DefaultImageSlug)
 }
 
 func testAccCheckCloudscaleServerConfig_update_state_running(rInt int) string {
@@ -410,11 +468,11 @@ func testAccCheckCloudscaleServerConfig_update_state_running(rInt int) string {
 resource "cloudscale_server" "basic" {
   name      					= "terraform-%d"
   flavor_slug    			= "flex-2"
-  image_slug     			= "debian-8"
+  image_slug     			= "%s"
   volume_size_gb			= 10
 	status 							= "running"
   ssh_keys = ["ecdsa-sha2-nistp256 AAAAE2VjZHNhLXNoYTItbmlzdHAyNTYAAAAIbmlzdHAyNTYAAABBBFEepRNW5hDct4AdJ8oYsb4lNP5E9XY5fnz3ZvgNCEv7m48+bhUjJXUPuamWix3zigp2lgJHC6SChI/okJ41GUY="]
-}`, rInt)
+}`, rInt, DefaultImageSlug)
 }
 
 func testAccCheckCloudscaleServerConfig_update_recreate(rInt int) string {
@@ -422,11 +480,11 @@ func testAccCheckCloudscaleServerConfig_update_recreate(rInt int) string {
 resource "cloudscale_server" "basic" {
   name      			    = "terraform-%d"
   flavor_slug    			= "flex-4"
-  image_slug     			= "debian-8"
+  image_slug     			= "%s"
   use_private_network		= true
   volume_size_gb			= 10
   ssh_keys 						= ["ecdsa-sha2-nistp256 AAAAE2VjZHNhLXNoYTItbmlzdHAyNTYAAAAIbmlzdHAyNTYAAABBBFEepRNW5hDct4AdJ8oYsb4lNP5E9XY5fnz3ZvgNCEv7m48+bhUjJXUPuamWix3zigp2lgJHC6SChI/okJ41GUY="]
-}`, rInt)
+}`, rInt, DefaultImageSlug)
 }
 
 func testAccCheckCloudscaleServerConfig_only_private_network(rInt int) string {
@@ -434,10 +492,10 @@ func testAccCheckCloudscaleServerConfig_only_private_network(rInt int) string {
 resource "cloudscale_server" "private" {
   name      			    = "terraform-%d"
   flavor_slug    			= "flex-4"
-  image_slug     			= "debian-8"
+  image_slug     			= "%s"
   use_private_network		= true
   use_public_network		= false
   volume_size_gb			= 10
   ssh_keys 						= ["ecdsa-sha2-nistp256 AAAAE2VjZHNhLXNoYTItbmlzdHAyNTYAAAAIbmlzdHAyNTYAAABBBFEepRNW5hDct4AdJ8oYsb4lNP5E9XY5fnz3ZvgNCEv7m48+bhUjJXUPuamWix3zigp2lgJHC6SChI/okJ41GUY="]
-}`, rInt)
+}`, rInt, DefaultImageSlug)
 }